move telemetry again
