+++ core/src/org/noble/helium/handling/TextureHandler.java
@@ -2,7 +2,7 @@ package org.noble.helium.handling;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
-import org.noble.helium.subsystems.telemetry.HeliumTelemetry;
+import org.noble.helium.subsystems.HeliumTelemetry;
 
 import java.util.HashMap;
 import java.util.Map;
