+++ core/src/org/noble/helium/lda/LDAExtractor.java
@@ -3,7 +3,7 @@ package org.noble.helium.lda;
 import com.badlogic.gdx.files.FileHandle;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
-import org.noble.helium.subsystems.telemetry.HeliumTelemetry;
+import org.noble.helium.subsystems.HeliumTelemetry;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
